spider man no way home

+++ internal/pkg/gatehttp/var.go
@@ -11,8 +11,8 @@ var TempHTML = template.New("html").Funcs(template.FuncMap{
 	"addFloatFloat": func(a, b float64) float64 {
 		return a + b
 	},
-	"mulFloatInt": func(a float64, b int) float64 {
-		return float64(a * float64(b))
+	"mulFloatInt": func(a float32, b uint32) float64 {
+		return float64(a * float32(b))
 	},
 })
